refact: move decryption code in utils/account to utils/decryption
