+++ apps/mobile/app/(authenticated)/(tabs)/(signer,explorer,converter)/signer/bitcoin/account/[id]/settings/export/descriptors.tsx
@@ -20,7 +20,6 @@ import { useBlockchainStore } from '@/store/blockchain'
 import { Colors } from '@/styles'
 import { type Secret } from '@/types/models/Account'
 import { type AccountSearchParams } from '@/types/navigation/searchParams'
-import { getAccountWithDecryptedKeys } from '@/utils/account'
 import {
   getExtendedKeyFromDescriptor,
   getFingerprintFromExtendedPublicKey
@@ -32,6 +31,7 @@ import {
   getMultisigScriptTypeFromScriptVersion,
   appNetworkToBdkNetwork
 } from '@/utils/bitcoin'
+import { getBitcoinWithDecryptedKeys } from '@/utils/decryption'
 import { shareFile } from '@/utils/filesystem'
 
 // Function to calculate checksum for descriptor using a simpler approach
@@ -90,7 +90,7 @@ export default function ExportDescriptors() {
         const isImportAddress =
           account.keys?.[0]?.creationType === 'importAddress'
 
-        const temporaryAccount = await getAccountWithDecryptedKeys(account)
+        const temporaryAccount = await getBitcoinWithDecryptedKeys(account)
 
         let descriptorString = ''
