+++ apps/mobile/utils/key.ts
